refactor(analytic-logs): Remove InteractionCountsRepository and related SQL migration
- Deleted InteractionCountsRepository and its associated SQL migration for interaction counts.
- Updated use cases to utilize WorkflowActivityCountsRepository for interaction trend and total interactions data retrieval.
- Adjusted shared module to remove InteractionCountsRepository references.
