+++ apps/api/src/app/shared/shared.module.ts
@@ -14,7 +14,6 @@ import {
   ExecuteBridgeRequest,
   featureFlagsService,
   GetDecryptedSecretKey,
-  InteractionCountsRepository,
   InvalidateCacheService,
   LoggerModule,
   ProviderVolumeCountsRepository,
@@ -127,7 +126,6 @@ const ANALYTICS_PROVIDERS = [
   WorkflowRunRepository,
   WorkflowActivityCountsRepository,
   ProviderVolumeCountsRepository,
-  InteractionCountsRepository,
   DeliveryTrendCountsRepository,
   WorkflowVolumeCountsRepository,
