--- libs/application-generic/src/services/analytic-logs/interaction-counts/index.ts
@@ -1,2 +1,0 @@
-export * from './interaction-counts.repository';
-export * from './interaction-counts.schema';
